+++ netpyne/simFuncs.py
@@ -727,13 +727,12 @@ def preRun():
     # reset all netstims so runs are always equivalent
     for cell in sim.net.cells:
         if cell.tags['cellModel'] == 'NetStim':
-            cell.hPointp.noiseFromRandom123(cell.gid, sim.id32('%d'%(cell.params['seed'])))
-            #cell.hRandom.Random123(cell.gid, sim.id32('%d'%(stim['seed'])))
-            #cell.hRandom.negexp(1)
+            cell.hRandom.Random123(cell.gid, sim.id32('%d'%(cell.params['seed'])))
+            cell.hRandom.negexp(1)
         for stim in cell.stims:
-            stim['hNetStim'].noiseFromRandom123(cell.gid, sim.id32('%d'%(stim['seed'])))
-            #stim['hRandom'].Random123(cell.gid, sim.id32('%d'%(stim['seed'])))
-            #stim['hRandom'].negexp(1)
+            if 'hRandom' in stim:
+                stim['hRandom'].Random123(cell.gid, sim.id32('%d'%(stim['seed'])))
+                stim['hRandom'].negexp(1)
 
 
 ###############################################################################
@@ -1094,7 +1093,7 @@ def saveData (include = None):
 
         # create folder if missing
         targetFolder = os.path.dirname(sim.cfg.filename)
-        if not os.path.exists(targetFolder):
+        if targetFolder and not os.path.exists(targetFolder):
             try:
                 os.mkdir(targetFolder)
             except OSError: